backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.7.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytz==2025.2
setuptools==80.4.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.6.0
